refactor facebook login
